Notification reason in exec_status msgs + fixed process search result

+++ ros2_bdi_skills/include/ros2_bdi_skills/bdi_action_executor.hpp
@@ -80,7 +80,7 @@ protected:
     float getProgress() {return progress_;} 
 
     /* Communicate execution status*/
-    void communicateExecStatus(const uint8_t& status);
+    void communicateExecStatus(const uint16_t& status);
 
     /*returns full action name*/
     std::string getFullActionName()
